Use unsigned instead of size_t

+++ mlir/include/mlir/Analysis/Matrix.h
@@ -15,58 +15,58 @@ template <typename INT> class Matrix {
 public:
   Matrix() = delete;
 
-  Matrix(size_t rows, size_t columns);
+  Matrix(unsigned rows, unsigned columns);
 
   Matrix(const Matrix &);
 
-  INT &operator()(size_t row, size_t column);
-  INT operator()(size_t row, size_t column) const;
+  INT &operator()(unsigned row, unsigned column);
+  INT operator()(unsigned row, unsigned column) const;
 
-  void swapColumns(size_t column, size_t otherColumn);
+  void swapColumns(unsigned column, unsigned otherColumn);
 
-  void swapRows(size_t row, size_t otherRow);
+  void swapRows(unsigned row, unsigned otherRow);
 
-  size_t getNRows() const;
+  unsigned getNRows() const;
 
-  size_t getNColumns() const;
+  unsigned getNColumns() const;
 
-  void resize(size_t newNRows, size_t newNColumns);
+  void resize(unsigned newNRows, unsigned newNColumns);
 
   friend std::ostream &operator<<(std::ostream &out, const Matrix &c);
 
 private:
-  size_t nColumns;
+  unsigned nColumns;
   std::vector<std::vector<INT>> data;
 };
 
 template <typename INT>
-Matrix<INT>::Matrix(size_t rows, size_t columns)
+Matrix<INT>::Matrix(unsigned rows, unsigned columns)
     : nColumns{columns}, data(rows, std::vector<INT>(columns, 0)) {}
 
 template <typename INT>
-INT &Matrix<INT>::operator()(size_t row, size_t column) {
+INT &Matrix<INT>::operator()(unsigned row, unsigned column) {
   assert(row < getNRows() && "Row outside of range");
   assert(column < getNColumns() && "Column outside of range");
   return data[row][column];
 }
 
 template <typename INT>
-INT Matrix<INT>::operator()(size_t row, size_t column) const {
+INT Matrix<INT>::operator()(unsigned row, unsigned column) const {
   assert(row < getNRows() && "Row outside of range");
   assert(column < getNColumns() && "Column outside of range");
   return data[row][column];
 }
 
-template <typename INT> size_t Matrix<INT>::getNRows() const {
+template <typename INT> unsigned Matrix<INT>::getNRows() const {
   return data.size();
 }
 
-template <typename INT> size_t Matrix<INT>::getNColumns() const {
+template <typename INT> unsigned Matrix<INT>::getNColumns() const {
   return nColumns;
 }
 
 template <typename INT>
-void Matrix<INT>::resize(size_t newNRows, size_t newNColumns) {
+void Matrix<INT>::resize(unsigned newNRows, unsigned newNColumns) {
   nColumns = newNColumns;
   for (auto &row : data)
     row.resize(nColumns, 0);
@@ -74,14 +74,14 @@ void Matrix<INT>::resize(size_t newNRows, size_t newNColumns) {
 }
 
 template <typename INT>
-void Matrix<INT>::swapRows(size_t row, size_t otherRow) {
+void Matrix<INT>::swapRows(unsigned row, unsigned otherRow) {
   assert(row < getNRows() && otherRow < getNRows()
                           && "Given row out of bounds");
   swap(data[row], data[otherRow]);
 }
 
 template <typename INT>
-void Matrix<INT>::swapColumns(size_t column, size_t otherColumn) {
+void Matrix<INT>::swapColumns(unsigned column, unsigned otherColumn) {
   assert(column < getNColumns() && otherColumn < getNColumns() &&
     "Given column out of bounds");
   for (auto &row : data)
@@ -93,14 +93,14 @@ std::ostream &operator<<(std::ostream &out, const Matrix<INT> &c) {
   out << "Dumping matrix, rows = " << c.getNRows()
       << ", columns: " << c.getNColumns() << '\n';
   out << "r/c  ";
-  for (size_t column = 0; column < c.getNColumns(); ++column) {
+  for (unsigned column = 0; column < c.getNColumns(); ++column) {
     out << "| " << std::setw(2) << column << " ";
   }
   out << std::endl;
   out << std::string(5 + c.getNColumns() * 5, '-') << std::endl;
-  for (size_t row = 0; row < c.getNRows(); ++row) {
+  for (unsigned row = 0; row < c.getNRows(); ++row) {
     out << std::setw(2) << row << " | ";
-    for (size_t column = 0; column < c.getNColumns(); ++column) {
+    for (unsigned column = 0; column < c.getNColumns(); ++column) {
       out << std::setw(4) << c(row, column) << " ";
     }
     out << std::endl;
