+++ mlir/unittests/Analysis/MatrixTest.cpp
@@ -6,13 +6,13 @@ namespace mlir {
 
 TEST(MatrixTest, ReadWrite) {
   Matrix<int> mat(5, 5);
-  for (size_t row = 0; row < 5; ++row) {
-    for (size_t col = 0; col < 5; ++col) {
+  for (unsigned row = 0; row < 5; ++row) {
+    for (unsigned col = 0; col < 5; ++col) {
       mat(row, col) = 10*row + col;
     }
   }
-  for (size_t row = 0; row < 5; ++row) {
-    for (size_t col = 0; col < 5; ++col) {
+  for (unsigned row = 0; row < 5; ++row) {
+    for (unsigned col = 0; col < 5; ++col) {
       EXPECT_EQ(mat(row, col), int(10*row + col));
     }
   }
@@ -20,14 +20,14 @@ TEST(MatrixTest, ReadWrite) {
 
 TEST(MatrixTest, SwapColumns) {
   Matrix<int> mat(5, 5);
-  for (size_t row = 0; row < 5; ++row) {
-    for (size_t col = 0; col < 5; ++col) {
+  for (unsigned row = 0; row < 5; ++row) {
+    for (unsigned col = 0; col < 5; ++col) {
       mat(row, col) = col == 3 ? 1 : 0;
     }
   }
   mat.swapColumns(3, 1);
-  for (size_t row = 0; row < 5; ++row) {
-    for (size_t col = 0; col < 5; ++col) {
+  for (unsigned row = 0; row < 5; ++row) {
+    for (unsigned col = 0; col < 5; ++col) {
       EXPECT_EQ(mat(row, col), col == 1 ? 1 : 0);
     }
   }
@@ -38,8 +38,8 @@ TEST(MatrixTest, SwapColumns) {
   mat.swapColumns(1, 3);
   mat.swapColumns(0, 4);
 
-  for (size_t row = 0; row < 5; ++row) {
-    for (size_t col = 0; col < 5; ++col) {
+  for (unsigned row = 0; row < 5; ++row) {
+    for (unsigned col = 0; col < 5; ++col) {
       EXPECT_EQ(mat(row, col), col == 1 ? 1 : 0);
     }
   }
@@ -47,14 +47,14 @@ TEST(MatrixTest, SwapColumns) {
 
 TEST(MatrixTest, SwapRows) {
   Matrix<int> mat(5, 5);
-  for (size_t row = 0; row < 5; ++row) {
-    for (size_t col = 0; col < 5; ++col) {
+  for (unsigned row = 0; row < 5; ++row) {
+    for (unsigned col = 0; col < 5; ++col) {
       mat(row, col) = row == 2 ? 1 : 0;
     }
   }
   mat.swapRows(2, 0);
-  for (size_t row = 0; row < 5; ++row) {
-    for (size_t col = 0; col < 5; ++col) {
+  for (unsigned row = 0; row < 5; ++row) {
+    for (unsigned col = 0; col < 5; ++col) {
       EXPECT_EQ(mat(row, col), row == 0 ? 1 : 0);
     }
   }
@@ -65,8 +65,8 @@ TEST(MatrixTest, SwapRows) {
   mat.swapRows(2, 0);
   mat.swapRows(0, 2);
 
-  for (size_t row = 0; row < 5; ++row) {
-    for (size_t col = 0; col < 5; ++col) {
+  for (unsigned row = 0; row < 5; ++row) {
+    for (unsigned col = 0; col < 5; ++col) {
       EXPECT_EQ(mat(row, col), row == 0 ? 1 : 0);
     }
   }
@@ -76,8 +76,8 @@ TEST(MatrixTest, Resize) {
   Matrix<int> mat(5, 5);
   EXPECT_EQ(mat.getNRows(), 5u);
   EXPECT_EQ(mat.getNColumns(), 5u);
-  for (size_t row = 0; row < 5; ++row) {
-    for (size_t col = 0; col < 5; ++col) {
+  for (unsigned row = 0; row < 5; ++row) {
+    for (unsigned col = 0; col < 5; ++col) {
       mat(row, col) = 10*row + col;
     }
   }
@@ -85,8 +85,8 @@ TEST(MatrixTest, Resize) {
   mat.resize(3, 3);
   EXPECT_EQ(mat.getNRows(), 3u);
   EXPECT_EQ(mat.getNColumns(), 3u);
-  for (size_t row = 0; row < 3; ++row) {
-    for (size_t col = 0; col < 3; ++col) {
+  for (unsigned row = 0; row < 3; ++row) {
+    for (unsigned col = 0; col < 3; ++col) {
       EXPECT_EQ(mat(row, col), int(10*row + col));
     }
   }
@@ -94,8 +94,8 @@ TEST(MatrixTest, Resize) {
   mat.resize(5, 5);
   EXPECT_EQ(mat.getNRows(), 5u);
   EXPECT_EQ(mat.getNColumns(), 5u);
-  for (size_t row = 0; row < 5; ++row) {
-    for (size_t col = 0; col < 5; ++col) {
+  for (unsigned row = 0; row < 5; ++row) {
+    for (unsigned col = 0; col < 5; ++col) {
       EXPECT_EQ(mat(row, col), row >= 3 || col >= 3 ? 0 : int(10*row + col));
     }
   }
